patients

+++ src/app/patient/routes.py
@@ -3,52 +3,56 @@ from ..models import db, Patient
 
 patient_bp = Blueprint('patient', __name__)
 
-
 @patient_bp.route('/patients', methods=['GET', 'POST'])
-def create():
+def manage_patients():
+    query = request.args.get('query')  # Obtener el término de búsqueda
+    
+    if query:
+        patients = Patient.query.filter(
+            (Patient.name.ilike(f'%{query}%')) |
+            (Patient.lastname.ilike(f'%{query}%')) |
+            (Patient.number.ilike(f'%{query}%')) |
+            (Patient.email.ilike(f'%{query}%'))
+        ).all()
+    else:
+        patients = Patient.query.all()
+
     if request.method == 'POST':
-        name = request.form['name']
-        lastname = request.form['lastname']
-        age = request.form['age']
-        contact = request.form['contact']
-        email = request.form['email']
-        country = request.form['country']
-        
-        patient = Patient(name=name, lastname=lastname, age=age, contact=contact, email=email, country=country, is_active='yes')
-        
-        db.session.add(patient)
-        db.session.commit()
+        # Crear nuevo paciente
+        if 'create' in request.form:
+            name = request.form['name']
+            lastname = request.form['lastname']
+            age = request.form['age']
+            number = request.form['number']
+            email = request.form['email']
+            country = request.form['country']
+            
+            new_patient = Patient(name=name, lastname=lastname, age=age, number=number, email=email, country=country, is_active='yes')
+            db.session.add(new_patient)
+            db.session.commit()
+            flash('Paciente creado exitosamente!', 'success')
+    
+        elif 'update' in request.form:
+            patient_id = request.form['patient_id']
+            patient = Patient.query.get(patient_id)
+            patient.name = request.form['name']
+            patient.lastname = request.form['lastname']
+            patient.age = request.form['age']
+            patient.number = request.form['number']
+            patient.email = request.form['email']
+            patient.country = request.form['country']
+            db.session.commit()
+            flash('Paciente actualizado exitosamente!', 'success')
 
-        flash('Patient created successfully!', 'success')
-        return redirect(url_for('patient.create'))
-    return render_template('/patients')
-
-@patient_bp.route('/patients/<int:id>', methods=['GET', 'POST'])
-def update(id):
-    patient = Patient.query.get(id)
-    if request.method == 'POST':
-        patient.name = request.form['name']
-        patient.lastname = request.form['lastname']
-        patient.age = request.form['age']
-        patient.contact = request.form['contact']
-        patient.email = request.form['email']
-        patient.country = request.form['country']
-        
-        db.session.commit()
-
-        flash('Patient updated successfully!', 'success')
-        return redirect(url_for('patient.update', id=id))
-    return render_template('/patients')
-
-@patient_bp.route('/patients/<int:id>', methods=['GET'])
-def delete(id):
-    patient = Patient.query.get(id)
-    db.session.delete(patient)
-    db.session.commit()
-    flash('Patient deleted successfully!', 'success')
-    return redirect(url_for('patient.delete'))
-
-@patient_bp.route('/patients', methods=['GET'])
-def get_all():
+        elif 'delete' in request.form:
+            patient_id = request.form['patient_id']
+            patient = Patient.query.get(patient_id)
+            db.session.delete(patient)
+            db.session.commit()
+            flash('Paciente eliminado exitosamente!', 'success')
+    
+    # Obtener lista de pacientes
     patients = Patient.query.all()
-    return render_template('/patients', patients=patients)
+    return render_template('patients.html', patients=patients)
+
+
